Add more tests to clear out state machine bugs
Timer-end callbacks (timerFired) is not passed as an event
to the rest of the state machine after NORMAL_STATE.
This reduces the amount of handling that has to be done
by each state.

For public functions, make them NOPs if it is not in a state that
allows such transition. giveUp() is not allowed after
"normal" phrase playing state.